Add cotp repo in the workspace (as submodule)
This is not the most efficient approach as we are
going to download and build a lot of stuff we do
not need, but is good enough to test variuos ideas.

Note that we will use a libify branch to push all
the changes required to make this work.
If it works we can then decide to make a proper
PR to the upstream repo (if they are interested)